+++ org.titou10.jtb.core/src/org/titou10/jtb/handler/SessionAddHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015 Denis Forveille titou10.titou10@gmail.com
+ * Copyright (C) 2025 Denis Forveille titou10.titou10@gmail.com
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -83,7 +83,8 @@ public class SessionAddHandler {
       newSessionDef.setHost2(dialog.getHost2());
       newSessionDef.setPort2(dialog.getPort2());
       newSessionDef.setHost3(dialog.getHost3());
-      newSessionDef.setPort3(dialog.getPort3());
+      newSessionDef.setHost3(dialog.getHost3());
+      newSessionDef.setDiscoveryFilter(dialog.getDiscoveryFilter());
 
       newSessionDef.setUserid(dialog.getUserId());
       newSessionDef.setActiveUserid(dialog.getUserId());