+++ org.titou10.jtb.qm.um/src/org/titou10/jtb/qm/um/UMQManager.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Denis Forveille titou10.titou10@gmail.com
+ * Copyright (C) 2025 Denis Forveille titou10.titou10@gmail.com
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -416,7 +416,7 @@ public class UMQManager extends QManager {
    }
 
    @Override
-   public boolean manulAcknoledge() {
+   public boolean manualAcknowledge() {
       return false;
    }